backend/requirements.txt
langchain
openai
PyPDF2
tiktoken
pinecone-client
"pinecone-client[grpc]"
python-dotenv